requirements.txt
django==1.9.5
django-solo
